rabbitmq-signadot-demo/consumer/requirements.txt
pika==1.3.0
requests==2.31.0

opentelemetry-instrumentation
opentelemetry-distro[otlp]==0.57b0